turned package builder into typescript file
